github-patches-check: allow all checks to run before failing
this check is migrated from that of net-next private:

>github-patches-check: allow all checks to run before failing
>
>Creating the environment to test commits in are expensive but the
>current commit check script terminates the job at the first error. This
>can be inefficient as multiple small errors may exists but only be
>revealed one at a time.
>
>Change the structure of how a commit it tested to allow multiple
>different checks to run before potentially failing the commit and halt
>the verification job. This allow all the tests to try and find as many
>faults as possible while still failing the commit if just a single check
>step fails.

+++ .github/deps/commits_check.sh
@@ -2,6 +2,20 @@
 
 set -e
 
+commit_status_init() {
+    COMMIT_STATUS_FAIL=0
+}
+
+commit_status_set_fail() {
+    local reason=$1
+    echo "Failed: $reason"
+    COMMIT_STATUS_FAIL=1
+}
+
+commit_status_get() {
+    echo $COMMIT_STATUS_FAIL
+}
+
 if [ -z "$1" ]; then
     echo "Please specify commit count."
     exit 1
@@ -12,6 +26,7 @@ git status > /dev/null
 
 for commit in $(git log --oneline --no-color -$1 --reverse | cut -d ' ' -f 1); do
     echo "============== Checking $commit ========================"
+    commit_status_init
 
     git checkout $commit
     PATCH_FILE=$(git format-patch -1 $commit)
@@ -38,18 +53,34 @@ for commit in $(git log --oneline --no-color -$1 --reverse | cut -d ' ' -f 1); d
     else
         echo $files
         # Run doc string checker on the files in the commit
-        ./kernel-doc -Werror -none $files
+        if ! ./kernel-doc -Werror -none $files; then
+            commit_status_set_fail "kernel-doc exited with non-zero return code"
+        fi
     fi
 
     echo
     echo "----------- Reverse xmas tree check ------------"
-    ./xmastree.py "$PATCH_FILE"
+    if ! ./xmastree.py "$PATCH_FILE"; then
+        commit_status_set_fail "xmastree.py exited with non-zero return code"
+    fi
 
     echo
     echo "----------- Checkpatch ------------"
-    ./utilities/checkpatch.py "$PATCH_FILE"
+    if ! ./utilities/checkpatch.py "$PATCH_FILE"; then
+        commit_status_set_fail "checkpatch.py exited with non-zero return code"
+    fi
 
     rm "$PATCH_FILE"
+
+    # Get the status of all checks and only move to next commit if all passed
+    echo
+    echo "----------- Checks summary --------------"
+    if [[ $(commit_status_get) -eq 0 ]]; then
+        echo "All checks passed for $commit"
+    else
+        echo "One or more checks failed for $commit"
+        exit 1
+    fi
 
     echo "========================================================"
     echo
